+++ py/tests/test_processing.py
@@ -252,7 +252,7 @@ def test_validate_cidr_condition():
 
 
 def test_validate_version_condition():
-    condition = '{"op": "version", "name": "event.release", "value": [">=1.2.0, <2.0.0", "~>3.1"]}'
+    condition = '{"op": "version", "name": "event.release", "value": [">=1.2.0, <2.0.0", "<1.0.0"]}'
     sentry_relay.validate_rule_condition(condition)